fix: Remove unnecessery additional Repository package and outdated files from Repository
