+++ Repository/ContratoRepository.java
@@ -1,4 +1,4 @@
-
+package Repository;
 
 import java.util.ArrayList;
