+++ Repository/GrupoConsorcioRepository.java
@@ -1,4 +1,4 @@
-
+package Repository;
 
 import Enums.StatusGrupoConsorcioEnum;
 import Models.GrupoConsorcio;
@@ -22,89 +22,79 @@ public class GrupoConsorcioRepository {
         }
     }
 
-    public void getGrupoByName(String nomeGrupo) {
-        GrupoConsorcio grupoTemp = null;
-
+    public GrupoConsorcio getGrupoByName(String nomeGrupo) {
         for (GrupoConsorcio grupo : grupos) {
             if (grupo.getNomeGrupo().equalsIgnoreCase(nomeGrupo)) {
-                grupoTemp = grupo;
+                System.out.println(grupo);
+                return grupo;
             }
         }
+        return null;
+    }
 
-        if (grupoTemp != null) {
-            System.out.println(grupoTemp);
-        } else {
-            System.out.println("\nO grupo procurado não foi encontrado!\n");
+    public GrupoConsorcio createGrupoConsorcio(String nomeGrupo, int numeroParticipantes, double valorTotal, double taxaAdmin) {
+        if(getGrupoByName(nomeGrupo) == null){
+            GrupoConsorcio grupoConsorcio = new GrupoConsorcio(nomeGrupo, numeroParticipantes, valorTotal, taxaAdmin);
+
+            grupos.add(grupoConsorcio);
+
+            System.out.println(grupoConsorcio);
+
+            System.out.printf("Grupo '%s' criado e adicionado com sucesso!\n\n", grupoConsorcio.getNomeGrupo());
+            return grupoConsorcio;
         }
+        return null;
     }
 
-    public void createGrupoConsorcio(String nomeGrupo, int numeroParticipantes, double valorTotal, double taxaAdmin) {
-        GrupoConsorcio grupoConsorcio = new GrupoConsorcio(nomeGrupo, numeroParticipantes, valorTotal, taxaAdmin);
-
-        grupos.add(grupoConsorcio);
-
-        System.out.println(grupoConsorcio);
-
-        System.out.printf("Grupo '%s' criado e adicionado com sucesso!\n\n", grupoConsorcio.getNomeGrupo());
-    }
-
-    public void updateGrupoConsorcio(
+    public GrupoConsorcio updateGrupoConsorcio(
             String nomeGrupo,
             int novoNumeroParticipantes,
             double novoValorTotal,
             double novoTaxaAdmin,
             StatusGrupoConsorcioEnum statusGrupoConsorcio
     ) {
-        GrupoConsorcio grupoTemp = null;
-
-        for (int i = 0; i < grupos.size(); i++) {
-            if (grupos.get(i).getNomeGrupo().equalsIgnoreCase(nomeGrupo)) {
-                if(novoNumeroParticipantes >= 0) {
-                    grupos.get(i).setNumeroParticipantes(novoNumeroParticipantes);
-                }
-                if(novoValorTotal >= 0) {
-                    grupos.get(i).setValorTotal(novoValorTotal);
-                }
-                if(novoTaxaAdmin >= 0) {
-                    grupos.get(i).setTaxaAdmin(novoTaxaAdmin);
-                }
-                if(statusGrupoConsorcio != null) {
-                    grupos.get(i).setStatusGrupoConsorcio(statusGrupoConsorcio);
-                }
-
-                double novoValorParcela = ((novoValorTotal + novoValorTotal * novoTaxaAdmin) / novoNumeroParticipantes);
-
-                if(novoValorParcela >= 0) {
-                    grupos.get(i).setValorParcela(novoValorParcela);
-                }
-
-                grupoTemp = grupos.get(i);
-            }
-        }
+        GrupoConsorcio grupoTemp = getGrupoByName(nomeGrupo);
 
         if (grupoTemp != null) {
-            System.out.println(grupoTemp);
+            if (novoNumeroParticipantes > 0) {
+                grupoTemp.setNumeroParticipantes(novoNumeroParticipantes);
+            }
 
-            System.out.printf("\nGrupo '%s' criado e adicionado com sucesso!\n", grupoTemp.getNomeGrupo());
+            if (novoValorTotal > 0) {
+                grupoTemp.setValorTotal(novoValorTotal);
+            }
+
+            if (novoTaxaAdmin >= 0) {
+                grupoTemp.setTaxaAdmin(novoTaxaAdmin);
+            }
+
+            if (statusGrupoConsorcio != null) {
+                grupoTemp.setStatusGrupoConsorcio(statusGrupoConsorcio);
+            }
+
+            double novoValorParcela = (novoValorTotal + (novoValorTotal * novoTaxaAdmin)) / novoNumeroParticipantes;
+
+            if (novoValorParcela > 0) {
+                grupoTemp.setValorParcela(novoValorParcela);
+            }
+
+            System.out.println(grupoTemp);
+            System.out.printf("\nGrupo '%s' atualizado com sucesso!\n", grupoTemp.getNomeGrupo());
+            return grupoTemp;
         } else {
-            System.out.println("Grupo não foi encontrado, logo não foi posspivel atualizá-lo!\n");
+            System.out.println("Erro: Grupo não encontrado para atualização!");
         }
+        return null;
     }
 
     public void deleteGrupoConsorcio(String nomeGrupo) {
-        GrupoConsorcio grupoTemp = null;
-        for (int i = 0; i < grupos.size(); i++) {
-            if (grupos.get(i).getNomeGrupo().equalsIgnoreCase(nomeGrupo)) {
-                grupoTemp = grupos.get(i);
-            }
-        }
+        GrupoConsorcio grupoTemp = getGrupoByName(nomeGrupo);
 
         if (grupoTemp != null) {
             grupos.remove(grupoTemp);
-
-            System.out.printf("Grupo: %s deletado com sucesso!\n", grupoTemp.getNomeGrupo());
+            System.out.printf("Grupo '%s' deletado com sucesso!\n", grupoTemp.getNomeGrupo());
         } else {
-            System.out.println("Não foi possível deletar o grupo, pois esse não foi encontrado!");
+            System.out.println("Grupo não encontrado para exclusão!");
         }
     }
 }
